Migrate TLS patches to new patching system.
