+++ src/core/cpu_patches.h
@@ -9,6 +9,7 @@ class CodeGenerator;
 
 namespace Core {
 
+/// Patches CPU instructions that cannot run as-is on the host.
 void PatchInstructions(u64 segment_addr, u64 segment_size, Xbyak::CodeGenerator& c);
 
 } // namespace Core
