+++ src/core/module.cpp
@@ -7,11 +7,10 @@
 #include "common/logging/log.h"
 #include "common/string_util.h"
 #include "core/aerolib/aerolib.h"
-#include "core/instruction_emulator.h"
+#include "core/cpu_patches.h"
 #include "core/loader/dwarf.h"
 #include "core/memory.h"
 #include "core/module.h"
-#include "core/tls.h"
 
 namespace Core {
 
@@ -132,7 +131,6 @@ void Module::LoadModuleToMemory(u32& max_tls_index) {
 
             add_segment(elf_pheader[i]);
             if (elf_pheader[i].p_flags & PF_EXEC) {
-                PatchTLS(segment_addr, segment_file_size, c);
                 PatchInstructions(segment_addr, segment_file_size, c);
             }
             break;
