+++ src/core/tls.h
@@ -22,13 +22,13 @@ struct Tcb {
     void* tcb_thread;
 };
 
+/// Gets the thread local storage key for the TCB block.
+u32 GetTcbKey();
+
 /// Sets the data pointer to the TCB block.
 void SetTcbBase(void* image_address);
 
 /// Retrieves Tcb structure for the calling thread.
 Tcb* GetTcbBase();
-
-/// Patches any instructions that access guest TLS to use provided storage.
-void PatchTLS(u64 segment_addr, u64 segment_size, Xbyak::CodeGenerator& c);
 
 } // namespace Core
